Move async to sync conversion to helpers
